chore: change module to modifier keyworks
